+++ src/main/client/client/GUI/Forms/Template.java
@@ -1,0 +1,18 @@
+package client.GUI.Forms;
+
+import client.GUI.CPSClientGUI;
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+
+public class Template {
+
+    @FXML
+    private Button btnBack;
+
+    @FXML
+    void returnToMain(ActionEvent event) {
+        CPSClientGUI.goBack(true);
+    }
+
+}