+++ src/main/server/server/CPSServer.java
@@ -4,7 +4,6 @@ import controller.Controllers;
 import ocsf.server.AbstractServer;
 import ocsf.server.ConnectionToClient;
 import org.apache.commons.cli.*;
-import static controller.Controllers.*;
 
 import java.io.IOException;
 import java.sql.SQLException;
@@ -171,7 +170,9 @@ public class CPSServer extends AbstractServer
         try
         {
             Controllers.initDb(dbUrl, dbUsername, dbPwd);
+            System.out.print("Please wait, setting up... ");
             Controllers.init();
+            System.out.println("Done!");
             initDummies();
         }
         catch (SQLException e)