+++ src/main/client/Changelog.md
@@ -1,7 +1,19 @@
 # 13.01.2018
-## Login Form
+## General
+* Handled error message on the client will only be printed if the --debug flag is active.
+
+## Login / Registration
 * Added "Password" field
-* Car selection is now a combobox in Enter Parking
 * Grouped and ordered FXML decleration
 * Upgraded "Session" object to suppert multiple types of users.
 * Fixed register/login flow to work with new implementations.
+
+## Enter Parking
+* Car selection is now a combobox 
+* This form now longer allows selection of parking lots. It will use the parking lot the user is currently logged into.
+
+## Manage Cars
+* Manage Cars screen now ready to rock.
+
+## View Preorders
+* Fixed the 'Refresh' functionality.